refactor(tui): share reload-resort test fixture across handler test modules
The setup_reload_resort_fixture helper was added inline in
detail_view_handlers.rs to drive the KAN-534 Step 1 tests. The same
fixture is needed by upcoming tests for the remaining offender sites in
popup_handlers, dialog_handlers, and app/mod.rs. Move it to a shared
crate::test_helpers module so each test module imports the same builder
instead of duplicating the seeding logic.

No behavior change: the five Step 1 tests still pass against the moved
fixture. The fixture now also stores board_id, column_id, b_id, and c_id
so upcoming tests that need to mutate the fixture (priority updates,
sprint assignments, etc.) have ids available without re-seeding.
